app/controller/pyproject.toml
[tool.poetry]
name = "resaas-re-job-controller"
version = "0.1.0"
description = "RESAAS component which, within a job received from the scheduler, manages optimization and final runs"
authors = ["simeoncarstens <simeon.carstens@tweag.io>"]
packages = [ { include = 'resaas' } ]



[tool.poetry.scripts]
resaas-controller = 'resaas.controller.run:run'
resaas-controller-local = 'resaas.controller.run_local:run'

[tool.poetry.dependencies]
python = "^3.8"
numpy = "*"
resaas-common = { path = "../../lib/common", develop = true }
resaas-grpc = { path = "../../lib/grpc", develop = true }
resaas-schedule-optimization = { path = "../../lib/schedule_estimation", develop = true }
click = "^7.1.2"

# Runners can be installed as extras for convenience
resaas-runner-rexfw = { path = "../../lib/runners/rexfw", optional = true, develop = true }
grpcio = "^1.35.0"
scipy = "^1.6.1"

[tool.poetry.extras]
rexfw = ["resaas-runner-rexfw"]

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
black = "^20.8b1"
flake8 = "^3.8.4"
pylint = "^2.6.0"
mypy = "^0.790"
pytest-dotenv = "^0.5.2"
pytest-cov = "^2.10.1"
isort = "^5.7.0"
jedi = "*"
rope = "*"
pyyaml = "*"
numpy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
env_override_existing_values = 1
env_files = [".env"]

[tool.coverage.run]
omit = ["*test*"]
